+++ fns-hl7-pipeline/fn-lake-segs-transformer/src/main/kotlin/gov/cdc/dex/hl7/Function.kt
@@ -12,7 +12,6 @@ import gov.cdc.dex.metadata.SummaryInfo
 import gov.cdc.dex.util.DateHelper.toIsoString
 import gov.cdc.dex.util.JsonHelper
 import gov.cdc.dex.util.JsonHelper.toJsonElement
-import gov.cdc.dex.util.ProcessingStatus.PSClientUtility
 import org.slf4j.LoggerFactory
 import java.lang.IllegalArgumentException
 import java.util.*
@@ -50,10 +49,6 @@ class Function {
         val outList = mutableListOf<String>()
         val profileFilePath = "/basicProfile.json"
         val config = listOf(profileFilePath)
-        val psClientUtility = PSClientUtility()
-        var traceId :String =""
-        var spanId :String =""
-        var childSpanId :String? = null
 
         messages.forEachIndexed { messageIndex: Int, singleMessage: String? ->
             try {
@@ -68,25 +63,12 @@ class Function {
                         filePath =
                             JsonHelper.getValueFromJson("routing_metadata.ingested_file_path", inputEvent).asString
                         messageUUID = JsonHelper.getValueFromJson("message_metadata.message_uuid", inputEvent).asString
-                        traceId = JsonHelper.getValueFromJson("routing_metadata.trace_id", inputEvent).asString
-                        spanId = JsonHelper.getValueFromJson("routing_metadata.span_id", inputEvent).asString
-
                         var status: String
                         var lakeSegsModel: List<Segment>?
                         var exception: Exception?
 
                         logger.info("DEX::Received and Processing messageUUID: $messageUUID, filePath: $filePath")
                         try {
-                            fnConfig.psURL?.let  {
-                                childSpanId =  psClientUtility.sendTraceToProcessingStatus(
-                                    fnConfig.psURL,
-                                    traceId,
-                                    spanId,
-                                    LakeSegsTransStageMetadata.PROCESS_NAME
-                                )
-                                logger.info("Span ID of messageUUID: $messageUUID : ${childSpanId}")
-                            }
-
                             // read the profile
                             val profile = this::class.java.getResource(profileFilePath)?.readText()
                                 ?: throw IllegalArgumentException("Unable to load profile $profileFilePath")
@@ -122,19 +104,6 @@ class Function {
                             exception = e, config = config, outList = outList
                         )
                     } // .try
-                    finally { // closing the span for the msg
-                        fnConfig.psURL?.let {
-                            childSpanId?.let {
-                                psClientUtility.stopTrace(
-                                    fnConfig.psURL,
-                                    traceId,
-                                    it,
-                                    LakeSegsTransStageMetadata.PROCESS_NAME
-                                )
-                            }
-                        }
-                    }
-
                 } // .if
             } catch (e : Exception) {
                 logger.error("ERROR: An unexpected error occurred: ${e.message}")
@@ -147,6 +116,7 @@ class Function {
             logger.info("Sent batch of ${outList.size} messages to ${fnConfig.evHubSendName}")
         } catch (e : Exception) {
             logger.error("Unable to send to event hub ${fnConfig.evHubSendName}: ${e.message}")
+            throw e
         }
 
         return outList