+++ fns-sidecars/hl7-file-sink/deployZip.sh
@@ -15,6 +15,7 @@ base_name=az-fun-$function_rootname.zip
 
 echo "Building Jar..."
 mvn clean package -DskipTests=true -Paz-$env
+rm base_name
 
 echo "Zipping it:"
 
@@ -24,12 +25,13 @@ zip -r ../../../$base_name *
 cd ../../..
 
 echo "Deploying Zip..."
-
+export LANG=C.UTF-8
 for i in "${arr[@]}"
 do
   echo "    Deploying $i on $1 Environment"
   az functionapp deployment source config-zip -g $hl7RG -n $function_rootname-$i-$env --src $base_name
     ### Set FN_VERSION:
+
   fn_version=$(cat pom.xml |grep -oPm1 "(?<=<version>)[^<]+")
   az functionapp config appsettings set --name $function_rootname-$i-$env --resource-group $hl7RG --settings FN_VERSION=$fn_version
 done